[4.0] com_contacts fields (#28275)
Contacts has an option to show custom user fields. I thought it wasn't working as the list didnt show any of my fields. On closer inspection the problem was that it is supposed to be "field groups" not "fields". This PR changes the field label from "User Custom Fields" to "User Custom Field Groups"