[4.0] Pagination
This PR corrects the markup in the pagination and as a result removes the need for the additional custom css. The onhover highlighting works a bit better but you probably wont notice that. We also get rounded corners as per the rest of the template.

Thanks @c-loddder and @sharkykz for pointing this out

As this changes the scss this pr requires you to run npm i or node build.js --compile-css or to download the prebuilt packages