[4.0] API Token
This PR corrects the markup in the xml to match the existing standards. It also removes some classes that are not required and fixes the switcher.

I also changed the name from Joomla Token to API Token as I believe it to be a more accurate description but maybe @nikosdion and @wilsonge disagree. I am not precious about the name

To test make sure you have enabled the user - token plugin

### before

### after