[4.0] send email to dotless domains
It is perfectly valid to have a dotless email address eg brian@example and all our validation rules allow it.

phpmailerdefaults to use PHP built-in FILTER_VALIDATE_EMAIL but this does not support dotless domain names

This PR removes an outdated comment and sets the validotr to be html5. This is the same validator used when submitting an email address and supports dotless domains.

The easiest way to test is with the global config

Set the from email to anything@example and then click the `send test mail` button

You will get an error message `Invalid address (to): anything@example`

Apply this PR and you no longer get the error