[3.x] send emails to dotless domains
Now that we are using phpmailer 6'+ this code block is not needed as it is now the default. However it does not support dotless domains which joomla supports so we still need to set a validator.

this is a backport of #28574

technically this pr is a b/c break as it allows dotless domains and it has not allowed allowed dotless domains since #24726 was merged but its much more of a bug fix as we should have allowed dotless domains and that pr broke it