[4.0] Tinymce column widths frontend
This PR resolves the release blocker #

It is a css change so you will need to npm i or node build.js --compile-css or use a prebuilt package from the download link below.

I believe that the root cause is related to the use of grid but if it is then I don't know how to fix that. Maybe someone else is able to.

### before

### after