[4.0] pull-right is not a bs4 class (#28764)
When you include an image you have the option to float the image left or right. this adds the class pull-left to the image. In BS4 we dont use pull- instead it is float-

Add an image to an article and select image float right

View on the front end

### before

### after