[4.0] th on safari
PR for #28904

This is a test PR for the TH alignment issue with safari. It is a bug in bootstrap that they have fixed and merged but not released yet. See https://github.com/twbs/bootstrap/pull/30323

This PR adds a _temporary_ reboot.css override with the code direct from bootstrap. I can't see any other way to fix this beta blocker.

I do not have access to Safari to test this.

@nikosdion please can you test this