Fix for exclude attribute in folderlist field (#22032)
Currently, the exclude attribute for the folderlist does not work. This is because the list of folders that are tested against the regular expression in exclude contains the full folder path and not the folder name only. So, instead of removing the root part after you test for the exclude, remove the root path before.