Fix API routing
Update htaccess.txt and web.config.txt

We add URL routing rules for the API application before the
existing rules for the public frontend. This way any URL which
starts with "/api" (case-sensitive) will always be redirected to
the Joomla API application instead of being diverted to the public
frontend.