[4.0] com_config buttons front end
Removes the toolbar markup as it is not needed - they are just buttons.
Bonus this also ensures the buttons are spaced correctly

To test. Install sample data and log in to the front end
Test the menu items for Site Settings and Template Settings and also test editing a module

Before this PR the buttons are next to each other without any spacing

After this PR the toolbar markup is removed and the buttons are spaced correctly