[4.0] Media Manager modal toolbar (#29681)
* [4.0] Media Manager modal toolbar

scss has been changed in this pr so you will need to node build.js --compile-css

### The problem
The toolbar buttons in the media manager modal eg when adding an intro image is displayed with the full colours and no spacing

### The solution
This could be technically wrong. It works but maybe there is a better way. It wraps the toolbar with the subhead div and adds a noshadow class

* Update administrator/components/com_media/tmpl/media/default.php

Co-authored-by: Quy <quy@fluxbb.org>

Co-authored-by: Quy <quy@fluxbb.org>