[4.0] Update help TOC and run
Thanks to @mbabker this PR replaces the existing script to generate the help toc.json file that is used to connect the help button in joomla to the help page on the docs site.

Other than a few comment changes it is the same code that @mbabker posted.

Before this PR then `php build\helpTOC.php` does not work

After applying this PR then it runs successfuly and generates the toc.json file

This PR includes an updated toc.json