[4.0] Discover extensions [a11y]
The select checkbox had two labels and the label for all checkboxes was identical.

To test make sure you have an extension listed as discoverable

The checkbox only has a label "Select" and the name of the extension is also a label.

After this PR the second label tag is removed and the first label says "Select <extension name">

See screenshots