[4.0] calendar field [a11y]
After reports from webaim and checking with Bruce Lawson it is not safe to rely on a "title" attribute to provide the required text for assistive technology.

In this PR the text from the title is now also present as sr-only text. The title is kept so that sighted users have a textual description of the button on hover