[4.0] Custom admin menu items alias
In the admin there are no SEF urls so there is no need to have a menu item alias for admin menu items. It just clutters the interface and confuses the user.

This PR _only_ removes the display of the alias it does not stop it being created.

To test create an admin menu.
Create a menu item in that menu