[4.0] user registration accessibility
The description strings were removed some time ago but they were not removed from the xml. This resulted in the fields being rendered with an aria-described being created that pointed to a non existent element.

### To test
enable user registration in com_users in the admin and then go to the user registration form in the front end

View the source of any of the inputs and you will see code like

Apply the pr and view the generated source again and the same input will look like