WebAuthn should check for either GMP or BCmath presence
Currently, Joomla 4 complains if GMP is not enabled which is wrong. BCmath will do just fine.

Note: we will be able to remove this check completely once [my upstream PR](https://github.com/fgrosse/PHPASN1/pull/80) to include a pure PHP fallback to BCmath is accepted and a new version of the `fgrosse/PHPASN1` library is released.