[4.0] Accessibility plugin position
This simple PR flips the display of the plugin from the right hand side to the left hand side in LTR languages and the opposite in RTL languages

This is to avoid clashing with the menu icon on mobile displays.

On desktops it probably makes more sense this way as its with the menus.

### Steps to reproduce the issue
Enable Plugin: System - Additional Accessibility Features
mobile view
scroll to bottom

### AFTER LTR

### AFTER RTL