[4.0] Numbers not digits (#30559)
I have checked this with my client Cambridge University Press and they have checked the usage within the Corpus. https://en.wikipedia.org/wiki/Cambridge_English_Corpus The corpus is a multi billion word database of the usage of the english language in the world. The corpus is about how language is used and not just about the dictionary meaning

They wrote

Interesting! I’ve had a look and it looks like ‘number’ is more common – there are only a small number of examples with ‘digit’.

It’s hard to know for sure, as there are only very few examples of the exact sentences you’ve given so I did a more general search and often ‘number’ is used in sentences like ‘they could steal your credit card number and password’. But, I had a scan through and there are definitely lots of examples of e.g. ‘use a password that contains letters, numbers and special characters’, so I think you’re right to go with your feeling on this.