[4.0] Display of horizontal mod_articles_news module (#30527)
* First step to make newsflash horizontal

* [4.0] Fix Alert double error display in Installation

* seperating multiple messages with border top

* Increase minimum length

* Fix meter displaying complete message prematurely

* Add strenthmeter attribute to installation

* Change meter values to display green when password requirements are met

* [4.0] {Cassiopea] Implementing Password length/meter

* [4.0] Fix Fieldset description display (#30435)

* [4.0] Url Language Code is not a prefix (#30440)

* [4.0] URL LANGUAGE CODE is not a prefix

* modified desc as suggested

* [4.0] Implementing display of fieldset descriptions for fieldset
children

* Taking off description

* scss cs

* use @Quy suggestion

* file cs

* [4.0] Media field fix (#30455)

* fix

* typo, (well nothing new)

* Allow media fields in subforms

- Cleanup Media field
- Expose initilazation method of Bootstrpa Modlas to a Joomla.Bootstrap object
- Initilise Modlas on the lifecycle of the Custom Element

* Paleolithic js mode only in the legacy…

* CS

* Proper check

* Pffff

* Update joomla-field-media.w-c.es6.js

* [4.0] Convert icon used for jgrid defaults to fas fa- (#30464)

* make icon-circle use fa-circle

* allows for including icon- in call

* move fa-fw to last class to fit convention (#30469)

* [4.0] Long labels wrapping (#30474)

When we have fields displayed in columns with the label displayed above the input there is a width limit of 240px on the label.

This results in long labels wrapping when it is not needed as the column is wider than 240px.

In english we dont see this is many places but the easiest is in the __configure edit screen_  fields

This simple PR changes the label width to 100% when the fields are displayed in columns like this.

As this is an scss change you will need to `node build.js --compile-css` in order to test

### Before

### After

* Fix mod_articles_latest (#30459)

* [4.0] Add a parameter for "back-to-top" button in Cassiopeia (#30441)

* [4.0] Add a parameter for back-to-top button in Cassiopeia

Added a parameter to show / hide the back-to-top button in Cassiopeia.
Currently the footer appears when a module has been published on position "footer". And the back-to-top only appears if the footer is visible. Since the footer is part of the grid, it is not easy to move the back-to-top outside the grid.

Now it is possible to decide to have a back-to-top or not. If yes, the footer will be rendered showing the button, independently if there is a module in the position footer.
It is also possible to show only modules (switch back-to-top off) or modules and back-to-top (switch on).

* Added smooth scrolling

Add smooth scrolling taking in account the preferences of the user (reduced motion).
Improve  language file.

* Correct height and position of back-to-top button

* Correct code style

* Correct code style 2

* Change smooth scrolling behavior to take in account reduced motion preferences

* Language file corrected, language string changed to match others

* Update language/en-GB/tpl_cassiopeia.ini

Link instead button

Co-authored-by: Brian Teeman <brian@teeman.net>

* Update templates/cassiopeia/scss/blocks/_global.scss

Comment for smooth scroll

Co-authored-by: Brian Teeman <brian@teeman.net>

Co-authored-by: Brian Teeman <brian@teeman.net>

* fix improper family name. (#30481)

* [4.0] update helptoc (#30490)

* [4.0] update helptoc

Removes the string for the expired cache page that no longer exists and removes it from the index displayed on the left hand side in the help page

* api

* [4.0] Add new permissions-policy to the HTTPHeaders Plugin (#30491)

* add permissions-policy

* add Permissions-Policy to the dropdown

* alpha order

* [4.0] Error when changing status of tagged content items (#30466)

* Correct column name

* Fix changing UCM state

* [4.0] br tag (#30503)

* [4.0] br tag

In Joomla 4 we use `<br>` not `<br />`

* bad grep

* [4.0] Remove old search component (#30506)

* [4.0] Remove old search component

The regular search component is not in J4

This pr removes reference to it in the help system
`administrator/index.php?option=com_admin&view=help`

* revert

* [4.0] Help Dashboard (#30508)

* [4.0] Help Dashboard

This PR makes multiple changes to the help dashboard. In order of importance they are.

- Created a new section "Start Here"
- Moved "Joomla Help" to the "Start Here" section to give it greater priority as it really is the first place you should look for help
- Removed target=_blank from "Joomla Help"
- Moved "Documentation Wiki" from "Resources" to the top of "Additional Help"

* deduplicate

* compass icon

* Moved css for mod_articles_news to media folder

User webAsset Manager to load specific css for module

* Move web asset call to horizontal.php

Co-authored-by: Jean-Marie Simonet <infografjms@gmail.com>
Co-authored-by: Quy <quy@fluxbb.org>
Co-authored-by: dGrammatiko <d.grammatiko@gmail.com>
Co-authored-by: Bear <programming@hallhome.us>
Co-authored-by: Brian Teeman <brian@teeman.net>
Co-authored-by: Christiane Maier-Stadtherr <dev@chmst.de>
Co-authored-by: Tobias Zulauf <zero-24@users.noreply.github.com>
Co-authored-by: SharkyKZ <sharkykz@gmail.com>