Create noTitle.php
This is for modules that generally do not have their header/title displayed, e.g. main menu or search. If used in the module position in the template index.php as "style=noTitle" instead of "style=none" then this will allow a mod id and module class suffix to be added to the modules.