[4.0] Admin menu redundant link
When adjacent links go to the same location this results in additional navigation and repetition for keyboard and screen reader users. WCAG 2.4.4

The parent menu item for Smart Search is the same as the first child  menu item.

Although the first one is disabled by the js and only acts as means to expand the submenu its still a problem for some assistive technologies.

This PR follows the pattern of all other core components and changes the parent link to index.php?option=com_finder

This can be tested either by performing a clean installation or by manually applying the sql in the update script.

Note that "database fix" will not apply the change in the update script as it is a content change not a structural change.