Create noTitle.php
This option can be used instead of "style-none" in the index.php to allow modules that generally do not have their header/title shown, e.g. main menu, search, but will add an id to the module - which can be used as anchors - and also include any module class suffix added to the module's setup advanced tab.