Fixing admin user activation url
If you have users activation approved by administrator and administrator is not logged in when he clicks the activation anchor in notification e-mail he will be redirected to the login page. Then after successful authorization he is redirected. The problem is that language filter changes the return url (the activation url) when it doesn't detect Itemid set in the url. With activation URL there is no Itemid set anywhere so it changes the return by wiping ANY redirection data like the most important part - the token. This patch just sets Itemid if it is an internal url and it was not set. That way any redirection data should be safe right now.