[4.0] Frontend Edit Article [a11y]
Every field must have a label. This PR ensures that a label is rendered for the text editor. It is not visually displayed as it has a class sr-only so there is no visible change.

If you apply the PR and inspect the source of the edit article page you will see that a label has been added as shown in the screenshot below

cc @carcam