[4.0] Article Info fields
In Joomla 4 we correctly changed the article info from a regular list to a definition. This has resulted in the items being indented and this is confusing some users as previously these items were not indented.

This PR removes the padding so that there is no visual change with J3 and yet maintains the correct semantic markup.

To test dont forget to rebuild the css