[4.0] Media manager icons
When the media manager is in list view it uses css data-type to generate an appropriate icon.

However the current code only matches lowercase file extensions. This PR make it case insensitive.

To test  upload an image with an uppercase filetype eg _DSC1234.JPG and go to the lust view and you will see there is no icon.

Apply this pr and run npm ci and refresh and now you have an icon