[4.0] Admin module checkout id [a11y] (#31608)
* [4.0] Admin module checkout id [a11y]

Every id on a page must be unique. Before this PR the id of the checked out icon in the admin modules Article Latest and Article Popular could be identical as it only used the number of the item in the list as the id. This PR additionally adds the name of the module and module id to ensure that the checked out id is unique.

Before the pr it will look like
`aria-labelledby="cbcheckin3-desc`

After the pr it will look like
`aria-labelledby="cbcheckinPopularArticles3-desc`