[4.0] Cassiopeia component.php (#31646)
The component.php view of the cassiopeia template does not have the code to use the font or colour schemes.

This can be seen by logging in to the site and opening an article to edit. Then select any of the xtd-editor plugins and you will see the missing buttons as in the screenshots below. Then apply the PR and you can see the bbuttons