[4.0] plugin installation order (#31489)
This PR does two things

The first is simply to put the list of plugins to be installed back into alphabetical order. This makes it easier to maintain and find things in the database

The second is to set the plugin order

In general it probably doesn't matter at all what order the plugins are in but if you look at the ordering dropdown it looks a mess as many of the plugins have the same order. Some of the oldest plugins have an ordering value but most just have a 0.

This PR adds a value starting at 1 for each plugin grouped by type

This is only for new installs
This is only for mysql (if approved I will update postgresql as well)