[4.0] Remove non working cutom help
In J3 there was the ability to add your own link (labeled custom help) on the help dropdown menu.

In J4 we no longer have a help dropdown - we have a dashboard instead which is constructed in a completely different way. In addition in Joomla 4 we have the ability to place a much more prominent custom support link as a module in the sidebar.

As a result this PR removes the code related to creating the non-working custom help link.

To test just make sure the admin menu works as it did before and that the remaining options in the admin menu module (show new and show help) still work.

Someone can always add this functionality back into Joomla at a later date - but it definitely wouldnt use this code as explained above and imho better to remove the non-working code and module options