[4.0] Skip-to plugin major upgrade [a11y]
This is a draft PR to give a heads up about this new version especially as there are new language strings.

Amongst other things it addresses @infograf768 comments about not everything be translatable and an old comment from @clodder about the dropdown script which is no longer being used.

In the process of creating this PR I discovered a few bugs in the upstream code (pr submitted) and that there github is ahead of the release on npm yet with a lower package number.

As a result this PR is not completed and its not working (hence its a draft).

This release is much much better under the hood. I have only implemented some of the features (ie not the important and all distinction) as they're more suitable to a content rich site than an admin ui.

For now you can only do a code review. As soon as I get a response to https://github.com/paypal/skipto/issues/68 I will be able to complete the release