[4.0] Password hint
A recent PR introduced the use of a placeholder as an indicator on the password field as a pseudo mask character.

The place holder text used was a string of *

However the browsers now use a bullet character and not an asterisk

This simple pr corrects that.

NOTE: because the symbol being emulated by this password placeholder is generated by the browser it is not possible to 100% match on all browsers. Images below are from chrome on windows 10