[4.0] skipto accessibility plugin major upgrade v4.0
SkipTo is a replacement for your old classic "Skipnav" link, (so please use it as such)! The SkipTo script creates a drop-down menu consisting of the links to important landmarks and headings on a given web page. The menu makes it easier for keyboard and screen reader users to quickly jump to the desired location by simply choosing it from the list of options.

Benefits
- All users can get an outline of the content on the page.
- Screen reader users can get a higher level navigation menu without having to use the screen reader landmark and header navigation commands which typically include longer lists of lower level headings and less used landmarks.
- Keyboard only users can more efficiently navigate to content on a page.
- Speech recognition users can use the menu to more efficiently navigate to content on a page.

How it works
- The SkipTo menu button is the first tabbable element on the page, and it is configured not to be visible when the page is loaded, the menu button becomes visible when it receives focus.
- Once the keyboard focus is on the menu button, pressing the ENTER or the SPACEBAR key will pull down the list of important landmarks and headings on the page.
- If you decide to reach the menu again, simply press the built-in access key alt+9

This plugin is enabled by default for the admin and can optionally be enabled for the frontend
This version addresses previous concerns about an additional dropdown menu js.
This version addresses previous concerns that the landmarks were not translatable.

To facilitate upgrades from previous beta releases I have added the list of removed files.

Thanks to the help of the people at https://github.com/paypal/skipto for adding some joomla specific changes