[4.0] Form label [a11y]
Each `<form>` element and form role that needs to be exposed as a landmark must be given an accessible name. This name will allow an assistive technology user to be able to quickly understand the purpose of the form landmark.

You can easily test this with the skipto plugin. Before this PR the item is just _Form_ After this PR the item is either _Form: New xxxx_ or _Form: Edit xxx_

- [x]com_categories
- [x]com_contact
- [x]com_content