[4.0] External link icon placement
As identified by @infograf768 the placement of the icon to indicate an external link is incorrect in RTL.

Unlike his proposal which creates additional css just for the RTL to wap the class from ::before to ::after this PR does it the correct way.

The icon is still ::before the beginning of the text and now correctly identifies the direction.

No need for a seperate class in the css at all. ;)