[4.0] Select checkbox [a11y]
Just as every link on a page should be unique so should a field.

On all the list views we have a checkbox to select the item.

On some lists they have a label (hidden) similar to "Select itemtitle" which ensures that the label is unique

This PR adds this to all the remaining list views that just have the label value of "select"

There is no visual change as the label has the class "visually hidden" so the best way to check is to view the generated source of a list view eg articles and featured articles

before this PR only articles checkboxes have a unique value eg Select Homepage
after this pr the featured articles list items also have a unique value