[4.0] svg logo [a11y]
Due to some "bugs" in some screenreaders and browsers and how they handle svg it is recommended to set them to focusable=false and ensure they have the correct role.

Because of the way we handle svg logos in the template this has to be done in the js and not the svg file itself.

The only way to really test after applying the PR and npm i is to view the source. If successable the svg will have the role and focusable attributes as below