[4.0] Smart Search Gather statistics
There are a few components that will give a warning message if a required plugin is not enabled AND the message includes a link directly to the plugin.

Smart Search -> Gather Statistics has a requirement for a specific option in the component to be enabled. This PR updates the message to provide a link to the options.