[4.0] Remove unused mixin
This mixin was aadded #27861

But then the only place it was used was removed during #31520

To test

make a copy of templates\cassiopeia\css\template.css
apply pr
regenerate the css (node build/build.js --compile-css)
compare the copy and the new template.css
The only difference should be for .container-header .mod-finder > label which is not using the correct code.

There should be no other differences.

(I really dont understand why we have so many bootstrap overrides that dont do anything)