[4.0] Web services (#32496)
* [4.0] Web services

This pr standardises the term to be "Web services". Note the capitalisation. This is inline with the naming used by the W3C.

The style guide rules of capitalising the words in a label still apply so other strings do not need to be updated.

I will submit an additional PR to style guide repo

* Update administrator/language/en-GB/com_config.ini

Co-authored-by: Quy <quy@fluxbb.org>

Co-authored-by: Quy <quy@fluxbb.org>