[4.0] CSP
Many many months ago we reached the point where no new features can be added to Joomla 4.0. We have to stop at some point and release Joomla 4.0 or it really will never be released and many many people's hard work will be a waste of time.

But at the same time there must be a point where we say that code that is not ready for release is removed from 4.0.

14 months ago we had that decision about com_workflows and the decision was to give a period of a few extra months or it would be removed possibly for a 4.1 release.

This PR removes com_csp. In its current form it is not usable as already reported and it is not documented as already reported except for some basic this button does this help. Considering the first commit of this component was almost THREE years ago it's fair to say that there is no evidence of any documentation any time soon.

I have NOT removed the httpheaders plugin as that might still be usable but might need a few tweaks after the component is removed as they appear to share some configuration options. But I wouldnt see anything wrong with removing that as well for the same reason and its already available as an extension and of course it is currently undocumented.

This component can be re-introduced for Joomla 4.1 if the bugs are resolved and documentation provided.