[4.0] subform buttons
Does it go somewhere? It's a link
Does it do something? It's a button

This PR corrects the usage of an `a` when it should be a `button` in the non-table repeatable layouts. You can see this in action in the CSP options after selecting custom mode and scrolling down the page.

There is no visible or functional change.

Note the javascript can probably be updated to not listen for a spacebar key as that is native for a button - this is beyond the scope of this PR.