[4.0] Repeatable grouped buttons [a11y]
The three buttons for add/delete/move that are displayed as a group for repeatable fields are too small. This is an accessibility and usability problem as it makes it difficult to select the correct button.

The apple ios guidelines for touch targets suggest 44px and the equivalent for android is 48.

By simply removing the btn-sm class from these buttons we satisfy those guidelines and its much much easieer to select the correct button