[4.0] Email Domain Options
The description of this option was on the repeatable subform. (See image 1) This is fine until you have some entries and the description is now lost at the bottom (see image 2). This PR moves the description to the fieldset which keeps it at the top and changes the styling so is much more consistent with similar tabs eg Mass Mail Users See image 3