[4.0] featured mobile
The number of cells in the table header should be the same as the number of cells in the table.

When workflow is enabled and you switch to mobile view then the status header is hidden but the content is not.

This PR makes sure the header is displayed at all sizes.

Any other comments about column width or what is displayed in the mobile view is beyond the scope of this pr and will be ignored.