[4.0] Atum badge borders
Some of the badges are set to have a border that is a color lighter than the badge text. Unfortunately that doesn't work if the text is white you get a white border.

All the badges need to have a visible border otherwise the badges will appear visually to be of different sizes.

This PR changes the border of the two existing white bordered badges to use the same shade of the dark text that the bg-info badge uses.

Testers dont forget to run npm to rebuild the css

The easies way to to check this would be in the article list after enabling the voting plugin