[4.0] table head font weight
Maybe its just me but I find it very unusual that the cells in the table head are not "bold" which is pretty normal.

It's not that obvious on a page like the article manager as all the cells are links but its really noticeable in the modules on the dashboard and in lists where not everything is a link eg access levels

On the assumption that its not just me here is a pr for consideration. As with all scss changes you will need to npm to build the css