[4.0] Admin modules bootstrap
Re-implements the missing code to output the bootstrap size and header tag for the admin modules.

At this time to test you will need to do a full install as there is NO update script yet (@ can you assist on that please)

There will be no visible difference but if you edit an admin module then you can change the bootstrap size and the header tag in the advanced tab of any admin module and it will have an effect. Before this pr they didnt do anything at all.